cdk/requirements.txt
aws-cdk-lib>=2.0.0
constructs>=10.0.0
aws-cdk.aws-apprunner-alpha>=2.0.0a1
